+++ search/delimiters.go
@@ -1,0 +1,14 @@
+package search
+
+import "github.com/launchdarkly/ld-find-code-refs/v2/options"
+
+func getDelimiters(opts options.Options) []string {
+	delims := []string{`"`, `'`, "`"}
+	if opts.Delimiters.DisableDefaults {
+		delims = []string{}
+	}
+
+	delims = append(delims, opts.Delimiters.Additional...)
+
+	return delims
+}
